requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
drf-spectacular>=0.29,<1.0
django-filter>=24.3,<25.0
python-dotenv>=1.2,<2.0
gunicorn>=22.0,<23.0
